Move persistent EMR files into one package
