+++ java/bulk-import/bulk-import-starter/src/main/java/sleeper/bulkimport/starter/executor/PlatformExecutor.java
@@ -20,7 +20,8 @@ import software.amazon.awssdk.services.emrserverless.EmrServerlessClient;
 import software.amazon.awssdk.services.sfn.SfnClient;
 import software.amazon.awssdk.services.sqs.SqsClient;
 
-import sleeper.bulkimport.starter.retry.ReturnBulkImportJobToQueue;
+import sleeper.bulkimport.starter.executor.persistent.PersistentEmrPlatformExecutor;
+import sleeper.bulkimport.starter.executor.persistent.ReturnBulkImportJobToQueue;
 import sleeper.core.properties.instance.InstanceProperties;
 import sleeper.core.properties.table.TablePropertiesProvider;
